+++ client/src/pages/Session.jsx
@@ -1,21 +1,40 @@
-import { useQuery } from '@apollo/client';
+import React from 'react';
 import Tabs from '../components/Tabs'
-import Auth from '../utils/auth';
-import M from 'materialize-css/dist/js/materialize.min.js';
-
-var elems = document.querySelectorAll('.tabs');
-var instance = M.Tabs.init(elems);
 
 const Session = () => {
-  return (
-    <main className='stndrd-page'>
-      <div>
-        <h5>Imaging Sessions</h5>
 
-        <Tabs/>
+    const tabs = [
+      {
+        id: 'tab1',
+        title: 'Tab 1',
+        content: <p>Content for Tab 1</p>,
+      },
+      {
+        id: 'tab2',
+        title: 'Tab 2',
+        content: <p>Content for Tab 2</p>,
+      },
+      {
+        id: 'tab3',
+        title: 'Tab 3',
+        content: <p>Content for Tab 3</p>,
+      },
+      {
+        id: 'tab4',
+        title: 'Tab 4',
+        content: <p>Content for Tab 4</p>,
+      },
+    ];
+  
+    return (
+      <main className='stndrd-page'>
+      <div className="container">
+      <h4>{user.username}s Imaging Journal</h4>
+        <h1>Materialize Tabs in React</h1>
+        <Tabs tabs={tabs} defaultTab="tab1" />
       </div>
-    </main>
-  );
-};
+      </main>
+    );
+  }
 
 export default Session;